Finished up CreateAccountActivity; Started work on UserAccoutActivity; Added functions to DatabaseHelper; Edited the database tables some; Created FunctionLibrary; Created UserSessionData;
Finished up CreateAccountActivity able to create a new account now. Started work on UserAccoutActivity where the user will be able to see and update their account. Added functions to DatabaseHelper to add new users, check if user exists, and to check username/password for logging in. Edited the database tables some to add columns that were missing and needed. Created FunctionLibrary to house functions that will be used in multiple activities like for error checking. Created UserSessionData to store the data of the currently logged in user.