Finished the UserReviewsActivity; Finished the LeaveReviewActivity; GUI adjustments to button sizes in most Activities; Started PackageDetailsActivity; Added functions to DatabaseHelper related to getting/setting/deleting user/users reviews;
UserReviewActivity is done. It has the ability to search by star count/rating, by year, or by month. Both ascending and descending and in order of date or star count/rating. User can also choose to edit or delete their own review and can filter the list so only their review is showing. All search fields can be used together or individually. Finished working on LeaveReviewActivity. User can  leave a review with or without text but a star rating has to be 1-5 in order to leave a review. GUI adjustments to button sizes in most Activities. Just resized them to be more in uniform with each other and to look a bit better. Started PackageDetailsActivity. About half way done with that. Just need to add more info and details on pricing.  Added functions to DatabaseHelper related to getting/setting/deleting user/users reviews. Along with adding more dummy data to get a larger pool of data to be used with sorting filters. Also created function to get the average star count/rating and display it in the UserReviewsActivity.