Finished the PackageDetailsActivity; Almost completed the first part of creating a new estimate(adding an address); Added some more error checking functions to the FunctionLibraryClass;
Finished the PackageDetailsActivity. I added a scrollView with information on the different packages offered in the app. Almost completed the first part of creating a new estimate(adding an address). I have the address able to be added to the database. Just need to add an edit and delete function. Added some more error checking functions to the FunctionLibraryClass.