Finished EnterAddressActivity; Created the remaining full CRUD functions on the user address; Create the PropertyMapActivity; Edited some function names;
Finished EnterAddressActivity GUI and added function for when "Edit Address" is clicked on the WelcomeUserActivity the current address data is loaded in to the edittext boxes. Created the remaining full CRUD functions on the user address.  Create the PropertyMapActivity, more yet to be done here. I'm currently testing the GoogleMaps on another project so no files get corrupted on the main project. Edited some function names to not be so "wordy".