Added GUI for the PropertyMapActivity; Setup the App to use GoogleMaps; Added map functionality;
Added GUI for the PropertyMapActivity to include buttons to interact with the GoogleMap. Setup the app to use GoogleMaps API adding dependencies to the gradle. Added map functionality that lets the user add points to a polygon creating an area to measure for acreage. I've got the area to work but on my test branch. Trying to get the adding of the polygon done first. Polygon tends to get weird once you add to many points. Also if you back track and try to add a point the polygon become unusable and overlaps itself. To fix this I'm trying to add draggable Markers to each point so as the marker moves so does that point. This way users can edit the polygon instead of trying to add more points. Still very much a work in progess.