Added Android Maps Utility Library; Added more dummy data; Worked on map functionality;
Added the Android Maps Utility Library to use with computing the area of the polygon. Added more dummy data to the SERVICE_ADDRESS and YARD_SERVICE tables. Worked on the map functionality. The user can now click on the map and set the points of the polygon. Then using a OnDragListener the user can move those points around adjust in the service area. While doing so the total acreage is displayed in a TextView for the user to see. The user can also undo their last point or rest the map completely. Just need to work on the database/save functionality to save the users data.