Finished EstOverviewActivity; Created math functions for setting totals; Create a custom map marker; Created Readme.txt;  Cleaned up the code; Created PackagePrice class; Error checked and fixed any found issues; Redid the statesList array;
Finished EstOverviewActivity. It now shows the address, property map with area of estimate, and total scores with selected packages. Created math functions for setting totals for the EstOverviewActivity. Totals include with and without taxes add. Prices are also adjusted to the total acreage. Created a custom map marker to match the color scheme of the app. Also made the marker a bit smaller. Created Readme.txt with a very brief description of the app.  Cleaned up the code a bit and organized some. Probably needs a lot more still. Created PackagePrice class to handle the package prices. Error checked and fixed any found issues though I'm sure there are still more to be found. Redid the statesList array so the whole state name is used instead of an abbreviation. This is the FINAL COMMIT for the project grade. Though I'll probably work on this more in the near future.